Authorization (Delete pet)

+++ src/actions/actions.ts
@@ -104,6 +104,14 @@ export async function editPet(petId: unknown, newPetData: unknown) {
 export async function deletePet(petId: unknown) {
   await sleep(1000);
 
+  // authentication check
+
+  const session = await auth();
+  if (!session) {
+    redirect("/login");
+  }
+
+  // validation
   const validatedPetId = petIdSchema.safeParse(petId);
 
   if (!validatedPetId.success) {
@@ -112,6 +120,29 @@ export async function deletePet(petId: unknown) {
     };
   }
 
+  // authentication check
+  const pet = await prisma.pet.findUnique({
+    where: {
+      id: validatedPetId.data,
+    },
+    select: {
+      userId: true,
+    },
+  });
+
+  if (!pet) {
+    return {
+      message: "Pet not found",
+    };
+  }
+
+  if (pet.userId !== session.user?.id) {
+    return {
+      message: "Not authorized.",
+    };
+  }
+
+  // database mutation
   try {
     await prisma.pet.delete({
       where: {
